refactor: address, location, place => address로 통일
