+++ src/components/common/input/labels.ts
@@ -6,7 +6,7 @@ export const LABELS = {
   NUMBER: '숫자',
   TITLE: '제목',
   PEOPLE: '인원',
-  LOCATION: '위치',
+  ADDRESS: '위치',
   TICKETS: '티켓수',
   LINK: '링크',
 } as const;
