+++ src/features/main/components/UpcomingPopupCard.tsx
@@ -1,7 +1,7 @@
 import CardComponent from '@/components/common/card';
 import useHandleSave from '@/hooks/useHandleSave';
 import type { Popup } from '@/types/popup';
-import { extractDistrict } from '@/utils/formatLocation';
+import { extractDistrict } from '@/utils/addressFormatter';
 
 interface UpcomingPopupCardProps {
   popup: Popup;
@@ -14,11 +14,11 @@ export default function UpcomingPopupCard({ popup, userId }: UpcomingPopupCardPr
   return (
     <CardComponent
       key={popup.id}
-      location={extractDistrict(popup?.address)}
+      {...popup}
+      address={extractDistrict(popup?.address)}
       // savedCount={popup.saveCount}
       isSaved={isSaved}
       onToggleSave={toggle}
-      {...popup}
       variant="compact"
     />
   );
