+++ src/features/popup-register/components/form-field/index.ts
@@ -1,9 +1,9 @@
+export { default as Address } from './Address';
 export { default as Capacity } from './Capacity';
 export { default as DateCalendar } from './DateCalendar';
 export { default as Description } from './Description';
 export { default as Email } from './Email';
 export { default as ExtraInfo } from './ExtraInfo';
-export { default as Location } from './Location';
 export { default as RecruitmentMethod } from './RecruitmentMethod';
 export { default as Tag } from './Tag';
 export { default as Thumbnail } from './Thumbnail';
