+++ src/features/popup-search/components/EventCard.tsx
@@ -3,6 +3,7 @@
 import CardComponent from '@/components/common/card';
 import useHandleSave from '@/hooks/useHandleSave';
 import type { PopupWithTags } from '@/types/popup';
+import { extractDistrict } from '@/utils/addressFormatter';
 
 interface EventCardProps {
   event: PopupWithTags;
@@ -12,5 +13,5 @@ interface EventCardProps {
 export default function EventCard({ event, userId }: EventCardProps) {
   const { isSaved, toggle } = useHandleSave(event.id, userId);
   console.log(event);
-  return <CardComponent {...event} isSaved={isSaved} onToggleSave={toggle} />;
+  return <CardComponent {...event} address={extractDistrict(event.address)} isSaved={isSaved} onToggleSave={toggle} />;
 }
